+++ src/common/helpers/profile.helper.ts
@@ -297,19 +297,3 @@ export const applyUserValues = (
 
   return result;
 };
-
-export const normalizeLayoutProperty = (layout?: PropertyLayout<string>) => {
-  if (!layout) return;
-
-  const normalizedLayout = { ...layout } as unknown as PropertyLayout<boolean>;
-
-  if (layout?.readOnly) {
-    normalizedLayout.readOnly = Boolean(layout?.readOnly);
-  }
-
-  if (layout?.isNew) {
-    normalizedLayout.isNew = Boolean(layout.isNew);
-  }
-
-  return normalizedLayout;
-};
