+++ src/common/services/recordToSchemaMapping/recordToSchemaMapping.service.ts
@@ -10,6 +10,7 @@ import { GRANDPARENT_ENTRY_PATH_INDEX } from '@common/constants/bibframe.constan
 import { ISelectedEntries } from '@common/services/selectedEntries/selectedEntries.interface';
 import { IUserValues } from '@common/services/userValues/userValues.interface';
 import { SchemaWithDuplicatesService } from '@common/services/schema';
+import { getParentEntryUuid } from '@common/helpers/schema.helper';
 
 // TODO: take into account a selected Profile
 export class RecordToSchemaMappingService {
@@ -167,7 +168,7 @@ export class RecordToSchemaMappingService {
       });
 
       // Parent schema entry has "block" type
-      if (this.updatedSchema?.get(entry.path[entry.path.length - 2])?.type === AdvancedFieldTypeEnum.block) {
+      if (this.updatedSchema?.get(getParentEntryUuid(entry.path))?.type === AdvancedFieldTypeEnum.block) {
         hasBlockParent = true;
       }
